--- a/argostranslate/argospm.py
+++ b/argostranslate/argospm.py
@@ -141,7 +141,7 @@ def main():
     remove_parser.add_argument("name", help="Package name")
 
     args = parser.parse_args()
-    if(hasattr(args, "callback")):
+    if hasattr(args, "callback"):
         args.callback(args)
     else:
         parser.print_help()