@yudelevi 2025-08-21
Reverted beam_size to >=4, results for some languages were horrible.